Move TestPublicKeyRepository to test fixtures
COAND-911
